rubocop

+++ app/controllers/items_controller.rb
@@ -36,9 +36,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if current_user.id == item.user_id
-      item.destroy
-    end
+    item.destroy if current_user.id == item.user_id
     redirect_to root_path
   end
 
@@ -54,10 +52,7 @@ class ItemsController < ApplicationController
   end
 
   def check_item_status
-    item_status = SalesInfo.find_by(item_id: params[:item_id])
-    if item_status != nil || current_user.id != @item.user_id
-      redirect_to root_path
-    end
+    item_status = SalesInfo.find_by(item_id: params[:id])
+    redirect_to root_path if !item_status.nil? || current_user.id != @item.user_id
   end
-
 end
